feat: LaundryItem Controller 세션 체크 추가(BE)

+++ clean4u/src/main/java/org/example/clean4u/laundryItem/LaundryItemController.java
@@ -1,7 +1,10 @@
 package org.example.clean4u.laundryItem;
 
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.example.clean4u._core.exception.Exception401;
+import org.example.clean4u.employee.Employee;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,7 +21,11 @@ public class LaundryItemController {
 
     // http://localhost:8080/laundry-item
     @GetMapping("/laundry-item")
-    public String laundryItemList(Model model) {
+    public String laundryItemList(Model model, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         List<LaundryItem> laundryItemList = repository.findAll();
         model.addAttribute("laundryItemList", laundryItemList);
         return "laundryItem/list-form";
@@ -26,7 +33,11 @@ public class LaundryItemController {
 
     // http://localhost:8080/laundry-item/{id}
     @GetMapping("/laundry-item/{id}")
-    public String detail(@PathVariable Long id, Model model) {
+    public String detail(@PathVariable Long id, Model model, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         LaundryItem laundryItem = repository.findById(id);
         model.addAttribute("laundryItem", laundryItem);
 
@@ -35,13 +46,21 @@ public class LaundryItemController {
 
     // http://localhost:8080/laundry-item/save
     @GetMapping("/laundry-item/save")
-    public String saveForm() {
+    public String saveForm(HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         return "laundryItem/save-form";
     }
 
     // http://localhost:8080/laundry-item/save
     @PostMapping("/laundry-item/save")
-    public String saveProc(@Valid LaundryItemRequest.SaveDTO saveDTO) {
+    public String saveProc(@Valid LaundryItemRequest.SaveDTO saveDTO, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         LaundryItem laundryItem = saveDTO.toEntity();
         repository.save(laundryItem);
         return "redirect:/laundry-item";
@@ -49,7 +68,11 @@ public class LaundryItemController {
 
     // http://localhost:8080/laundry-item/{id}/update
     @GetMapping("/laundry-item/{id}/update")
-    public String updateForm(@PathVariable Long id, Model model) {
+    public String updateForm(@PathVariable Long id, Model model, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         LaundryItem laundryItem = repository.findById(id);
         model.addAttribute("laundryItem", laundryItem);
 
@@ -67,14 +90,22 @@ public class LaundryItemController {
 
     // http://localhost:8080/laundry-item/{id}/update
     @PostMapping("/laundry-item/{id}/update")
-    public String updateProc(@PathVariable Long id, @Valid LaundryItemRequest.UpdateDTO updateDTO) {
+    public String updateProc(@PathVariable Long id, @Valid LaundryItemRequest.UpdateDTO updateDTO, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         repository.updateById(id, updateDTO);
         return "redirect:/laundry-item";
     }
 
     // http://localhost:8080/laundry-item/{id}/delete
     @PostMapping("/laundry-item/{id}/delete")
-    public String delete(@PathVariable Long id) {
+    public String delete(@PathVariable Long id, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         repository.deleteById(id);
         return "redirect:/laundry-item";
     }
